requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
jinja2==3.1.4
PyJWT==2.8.0
pydantic[email]==2.7.1
pandas==2.2.2
openpyxl==3.1.2
python-multipart==0.0.9
python-dotenv==1.0.1
aiosmtplib==3.0.1
